fix: enforce JWT auth on protected routes and repair order creation
  Two blocking problems found by running the stack against a real database,
  plus a small refactor. Both bugs predate this branch's endpoint work.

  1. POST /orders returned 500 - order creation never worked

  The CreateOrder query reused  in two incompatible positions: as the
  quantity column (integer) and as the right operand of a numeric
  multiplication, where Postgres infers numeric. It cannot deduce one type
  for both, so it rejected the statement at prepare time:

    pq: inconsistent types deduced for parameter

  sqlc generates and compiles this happily - only a live database rejects
  it, which is why the build stayed green while the feature was dead.
  Casting the reuse to the column's own type fixes it:

    VALUES (, , , , , ::numeric * ::integer)

  Deduced parameter types are unchanged ({integer,uuid,uuid,integer,
  numeric}), so the regenerated orders.sql.go differs only in its embedded
  SQL string - CreateOrderParams and the function signature are identical.

  2. Authentication was not enforced anywhere

  Two independent holes that compounded:

  - Authenticate is optional-auth by design: every failure path (no token,
    invalid token, unparseable subject) calls next.ServeHTTP and continues.
    It attaches a user, it never rejects. The router used it as though it
    enforced, so the authenticated route group was not.

  - getUserID read the X-USER-ID header, discarding the verified user the
    middleware had just put in the context. Any caller could claim any uuid,
    which defeated every ownership and actor check in the services - the
    order FSM's 403 guards included. A state machine can enforce state
    perfectly and still enforce identity not at all.

  The int64/uuid mismatch that originally forced the header placeholder was
  already fixed in c6f9004, so the placeholder could simply go:

  - add mw.RequiredAuth, which 401s unless Authenticate attached a user
  - getUserID now reads auth.UserFromContext; every call site is unchanged
  - split the router: browse endpoints public, everything else behind
    RequiredAuth (deny by default, so a forgotten route 401s rather than
    being silently public)

  This also delivers the intended browse-anonymously, authenticate-before-
  checkout flow.

  3. Move shared handler helpers to common.go

  getUserID, parseIDParam and statusFromServiceError lived in listing.go but
  are used by order.go and handler_auth.go too. Moved to their own file,
  mirroring response.go. No behaviour change.

  Verified against Postgres 16 with 70 assertions over the full lifecycle:
  public browse 200 without a token; orders and listing mutations 401
  without one; an X-USER-ID spoof with no token 401; pending -> confirmed ->
  paid -> completed; every actor guard 403; every wrong-state and
  terminal-state move 409; stock reserved on create and restored on cancel;
  strangers 403; missing orders 404; validation 400/409/404.

+++ backend/sql/queries/orders.sql
@@ -1,6 +1,6 @@
 -- name: CreateOrder :one
 INSERT INTO orders (listing_id, buyer_id, seller_id, quantity, unit_price, total_price)
-VALUES ($1, $2, $3, $4, $5, $5::numeric * $4)
+VALUES ($1, $2, $3, $4, $5, $5::numeric * $4::integer)
 RETURNING *;
 
 -- name: GetOrder :one
